split labeling and training seperatly
